pivony/server/requirements.txt
adal==1.2.7
altgraph==0.17.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
cx-Freeze==6.11.0.dev1
dnspython==2.2.1
Flask==2.1.2
Flask-Cors==3.0.10
gunicorn==20.0.4
idna==3.3
importlib-metadata==4.11.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
macholib==1.15.2
MarkupSafe==2.1.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
Nuitka @ https://github.com/Nuitka/Nuitka/archive/develop.zip
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.1
portalocker==2.4.0
pycparser==2.21
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.8
Werkzeug==2.1.2
zip-files==0.4.1
zipp==3.7.0
